working on setup

+++ src/process.cpp
@@ -16,7 +16,8 @@ namespace {
 	}
 }
 
-std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path) {
+std::unique_ptr<sdb::process> sdb::process::launch(
+	std::filesystem::path path, bool debug) {
 	pipe channel(/*close_on_exec=*/true);
 	pid_t pid;
 	if ((pid = fork()) < 0) {
@@ -25,7 +26,7 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path) {
 
 	if (pid == 0) {
 		channel.close_read();
-		if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+		if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
 			exit_with_perror(channel, "Tracing failed");
 		}
 		if (execlp(path.c_str(), path.c_str(), nullptr) < 0) {
@@ -37,14 +38,17 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path) {
 	auto data = channel.read();
 	channel.close_read();
 
-	if data.size() > 0 {
+	if (data.size() > 0) {
 		waitpid(pid, nullptr, 0);
 		auto chars = reinterpret_cast<char*>(data.data());
 		error::send(std::string(chars, chars + data.size()));
 	}
 
-	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end_=*/true));
-	proc->wait_on_signal();
+	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end_=*/true, debug));
+
+	if (debug) {
+		proc->wait_on_signal();
+	}
 
 	return proc;
 }
@@ -58,7 +62,7 @@ std::unique_ptr<sdb::process> sdb::process::attach(
 		error::send_errno("Could not attach");
 	}
 
-	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end_=*/false));
+	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/false, /*attached=*/true));
 	proc->wait_on_signal();
 
 	return proc;
@@ -67,12 +71,14 @@ std::unique_ptr<sdb::process> sdb::process::attach(
 sdb::process::~process() {
 	if (pid_ != 0) {
 		int status;
-		if (state_ == process_state::running) {
-			kill(pid_, SIGSTOP);
-			waitpid(pid_, &status, 0);
+		if (is_attached_) {
+			if (state_ == process_state::running) {
+				kill(pid_, SIGSTOP);
+				waitpid(pid_, &status, 0);
+			}
+			ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+			kill(pid_, SIGCONT);
 		}
-		ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-		kill(pid_, SIGCONT);
 
 		if (terminate_on_end_) {
 			kill(pid_, SIGKILL);
